third_party/WebKit/LayoutTests/fast/css/xml-stylesheet-alternate-no-title-expected.txt

xml alternate stylesheet with no title test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS getComputedStyle(element).getPropertyValue('color') is 'rgb(0, 128, 0)'
This text should be green
